Debugged issues with OpenAI connection and refactored file names
