--- open-fin-al/src/Gateway/AI/Model/IKeylessModel.ts
@@ -1,5 +1,0 @@
-import {IModel} from "./IModel";
-
-export interface IKeylessModel extends IModel {
-    create(model: string, messages: any[]) : Promise<any>;
-}
