+++ open-fin-al/src/Gateway/AI/Model/IKeylessModelGateway.ts
@@ -1,0 +1,5 @@
+import {IModelGateway} from "./IModelGateway";
+
+export interface IKeylessModelGateway extends IModelGateway {
+    create(model: string, messages: any[]) : Promise<any>;
+}
