+++ open-fin-al/src/Gateway/AI/Model/IModelGateway.ts
@@ -1,4 +1,4 @@
-export interface IModel {
+export interface IModelGateway {
     key?: any;
     create(model: string, messages: any[]) : Promise<any>;
 }
